// src/config/constants/types.ts
export interface Drip {
  id: string
  isActive: boolean
  address: string
  //lpAddress: string
  name: string
  subTitle: string
  description: string
  //logo: string
  tokenDecimals: number
  tokenSymbol: string
  projectSiteURL: string
  //dripStartBlock: number

  // pancake swap crap, remove later, breaks without it for now
  launchDate: string
  launchTime: string
  saleAmount: string
  raiseAmount: string
  cakeToBurn: string
  projectSiteUrl: string
  currency: string
  currencyAddress: string
  releaseBlockNumber: number
  campaignId?: string
  // pancake swap crap, remove later, breaks without it for now
}